% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changelog.R
\name{changelog}
\alias{changelog}
\title{Show changelog of a particular release}
\usage{
changelog(release = NULL, new_release = FALSE)
}
\arguments{
\item{release}{the result of \code{find_release} or NULL to show the current dev changelog.}

\item{new_release}{is it the first release?}
}
\description{
Show changelog of a particular release
}
