% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changelog.R
\name{changelog_between}
\alias{changelog_between}
\title{Show the closed github issues and merged pull requests between commits.}
\usage{
changelog_between(ref)
}
\arguments{
\item{ref}{character, revision range}
}
\description{
Show the closed github issues and merged pull requests between commits.
}
