% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagbot.R
\name{find_commit_to_tag}
\alias{find_commit_to_tag}
\title{find a commit to tag}
\usage{
find_commit_to_tag(version = NULL)
}
\arguments{
\item{version}{a character in semetic version. Use the latest CRAN release version if \code{NULL}.}
}
\description{
find a commit to tag
}
