% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{find_release}
\alias{find_release}
\title{Find a release that matches the CRAN release.}
\usage{
find_release(version = NULL)
}
\arguments{
\item{version}{the version to find. Use \code{NULL} to find the latest release version.}
}
\description{
Find a release that matches the CRAN release.
}
