% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagbot.R
\name{publish_release}
\alias{publish_release}
\title{Publish a github release based on CRAN or CRAN-like releases}
\usage{
publish_release(version = NULL)
}
\arguments{
\item{version}{the version to publish. Use \code{NULL} to publish the latest release.}
}
\description{
Publish a github release based on CRAN or CRAN-like releases
}
