% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagbot.R
\docType{package}
\name{tagbot-package}
\title{tagbot: Publish GitHub Releases}
\description{
Simplify the process of making a GitHub release. There are three major features provided by 'tagbot'. 1. It performs 'diff' between the source file of a release and the package's git repo to decide which commit is corresponding to the release. 2. It generates a changelog of closed issues and merged pull requested from GitHub API. 3. It publishes a GitHub release with the changelog.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RTagBot/tagbot}
  \item Report bugs at \url{https://github.com/RTagBot/tagbot/issues}
}

}
\author{
\strong{Maintainer}: Randy Lai \email{randy.cs.lai@gmail.com}

}
