% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagbot.R
\name{tagbot}
\alias{tagbot}
\title{Automatically create a github release from CRAN release}
\usage{
tagbot(version = NULL)
}
\arguments{
\item{version}{the version to search. Use \code{NULL} to search the latest release version.}
}
\description{
Automatically create a github release from CRAN release
}
