% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changelog.R
\name{changelog_between}
\alias{changelog_between}
\title{Show the closed github issues and merged pull requests between commits.}
\usage{
changelog_between(after, before = "HEAD")
}
\arguments{
\item{after}{character, the start commit}

\item{before}{character, the end commit}
}
\description{
Show the closed github issues and merged pull requests between commits.
}
