% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagbot.R
\name{find_release}
\alias{find_release}
\title{find a release with a given version}
\usage{
find_release(version = NULL)
}
\arguments{
\item{version}{the version string. Use the latest release version if \code{NULL}.}
}
\description{
find a release with a given version
}
