% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagbot.R
\docType{package}
\name{tagbot-package}
\title{tagbot: Publish GitHub Releases}
\description{
Simplify the process of making a GitHub release. There are three major features provided by 'tagbot'. First, it determines which commit corresponds to a release by diffing the source files of the release and the package's git repo. It also generates a changelog of closed issues and merged pull requests from GitHub API. And finally, it publishes a GitHub release with the changelog.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RTagBot/tagbot}
  \item Report bugs at \url{https://github.com/RTagBot/tagbot/issues}
}

}
\author{
\strong{Maintainer}: Randy Lai \email{randy.cs.lai@gmail.com}

}
