% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagbot.R
\docType{package}
\name{tagbot-package}
\title{tagbot: Git Tag Helper for CRAN Releases}
\description{
Identify which commit corresponds to a specific cran release.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RTagBot/tagbot}
  \item Report bugs at \url{https://github.com/RTagBot/tagbot/issues}
}

}
\author{
\strong{Maintainer}: Randy Lai \email{randy.cs.lai@gmail.com}

}
